Reduce difference with upstream
